Fix accounts being rendered multiple times
This fixes the bug where accounts in the sidebar are being rendered
multiple times whenever a modal is being opened.
