+++ lib/terrible_web/components/layouts/book.html.heex
@@ -1,40 +1,8 @@
 <div class="h-full">
-  <!-- Off-canvas menu for mobile, show/hide based on off-canvas menu state. -->
-  <div id="sidebar-mobile" class="relative z-50 lg:hidden" role="dialog" aria-modal="true">
-    <div class="fixed inset-0 bg-gray-900/80"></div>
-
-    <div class="fixed inset-0 flex">
-      <div class="relative mr-16 flex w-full max-w-xs flex-1">
-        <div class="absolute left-full top-0 flex w-16 justify-center pt-5">
-          <button type="button" class="-m-2.5 p-2.5">
-            <span class="sr-only">Close sidebar</span>
-            <svg
-              class="h-6 w-6 text-white"
-              fill="none"
-              viewBox="0 0 24 24"
-              stroke-width="1.5"
-              stroke="currentColor"
-              aria-hidden="true"
-            >
-              <path stroke-linecap="round" stroke-linejoin="round" d="M6 18L18 6M6 6l12 12" />
-            </svg>
-          </button>
-        </div>
-        <.sidebar
-          active_nav_item={@active_nav_item}
-          book={@book}
-          budget={@budget}
-          accounts={@streams.accounts}
-        />
-      </div>
-    </div>
-  </div>
-  <!-- Static sidebar for desktop -->
-  <div
-    id="sidebar-desktop"
-    class="hidden lg:fixed lg:inset-y-0 lg:z-50 lg:flex lg:w-72 lg:flex-col"
-  >
-    <.sidebar
+  <div class="hidden lg:fixed lg:inset-y-0 lg:z-50 lg:flex lg:w-72 lg:flex-col">
+    <.live_component
+      module={TerribleWeb.SidebarComponent}
+      id="sidebar-desktop"
       active_nav_item={@active_nav_item}
       book={@book}
       budget={@budget}
